Refactor onboarding flow to separate files
